+++ katana_public_api_client/api/sales_order/delete_sales_order.py
@@ -41,6 +41,11 @@ def _parse_response(
 
         return response_404
 
+    if response.status_code == 412:
+        response_412 = ErrorResponse.from_dict(response.json())
+
+        return response_412
+
     if response.status_code == 429:
         response_429 = ErrorResponse.from_dict(response.json())